backend/requirements.txt
flask
flask_sqlalchemy
flask_login
flask_cors
flask_bcrypt
flask_migrate
flask-limiter
docker
PyJWT>=2.6.0
bleach
gunicorn
bs4
psycopg2-binary